pyproject.toml
[project]
name = "slide-extractor"
version = "0.1.0"
description = "Lecture slide extraction pipeline using Meta SAM 3.1"
readme = "README.md"
authors = [
    { name = "LevMuchnik", email = "levmuchnik@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "torch>=2.7.0",
    "torchvision",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.22.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.20.0",
]

[project.scripts]
slide-extractor = "slide_extractor.cli:app"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu126"]
